+++ shared/src/test/scala/OptionFieldTests.scala
@@ -6,7 +6,8 @@ class OptionFieldTests
     extends FlatSpec
     with FormatTests {
 
-  case class Opt(x: Option[Int])
+  sealed trait Super
+  case class Opt(x: Option[Int]) extends Super
 
   object HideNull extends DerivedJsonProtocol {
     override def printNull = false
@@ -48,6 +49,27 @@ class OptionFieldTests
     import ShowNull._
     assert("{}".parseJson.convertTo[Opt] == Opt(None))
   }
+
+   {
+     import ShowNull._
+     implicit val superFmt = jsonFormat[Super]
+     "Option fields of ADTs" should behave like checkRoundtrip(
+       Opt(Some(2)): Super,
+       """{ "@type": "Opt", "x":2}"""
+     )
+   }
+
+  sealed trait Enum
+  case class Value(x: Int) extends Enum
+  case class Wrapper(enum: Option[Enum])
+
+  import ShowNull._
+  implicit val enumFormat: RootJsonFormat[Enum] = jsonFormat[Enum]
+  implicit val superFmt: RootJsonFormat[Wrapper] = jsonFormat[Wrapper]
+  "Option fields of inner ADTs" should behave like checkRoundtrip(
+    Wrapper(Some(Value(1))),
+    """{"enum":{"@type":"Value", "x": 1}}"""
+  )
 
 }